Fix format on empty has_one/has_many relationships
Before these changes, empty has_one/has_many relationships were
formatting as either {“type”: “author”, id: “”} with a blank ID for a
has_one or an empty object ({}) for a has_many.

JSON API spec requires null instead of a resource identifier object for
an empty has_one relationship, and an empty array ([]) for an empty
has_many relationship.

This change ensures that the `:data` key is formatted as nil/[] when
appropriate.