Format query params for pagination links based on config format
This PR formats the query params in pagination links to be consistent
with the configured `key_format`. It also will format dasherized query
params to underscores during deserialization.

Note: Because the query params were always underscored and because the
default `key_format` is dasherized, this may cause issues for people
that were relying on this bug (Eg. they expect underscored query params
even though `key_format` is configured to `:dasherized`).