Improve deprecation messages
The deprecations logged are inconsistent and don't contain enough
information to quickly identify the offending code. This change makes
them consistent in how they are displayed and also includes stacktrace
information in the deprecation, which makes it simple to find the
source.

Previously some deprecations forced ANSI coloring while others had no
coloring at all. This change makes them all colored consistently but
only applies the coloring if elixir detects that it is supported.

To get a useful stacktrace out of `normalize_relation_opts`, I had to
pass the `__CALLER__` in from the macro to that module function so it
could be used in the deprecation. There was seemingly no way to get at
the information otherwise.