Changes
1. Added `bare: true` to allow `JaSerializer.Builder.ResourceObject` to
build a bare object without requiring relationships. This can probably
be expected to `meta` and `links` as well. TBD.
2. Changes `JaSerializer.Formatter` to always output all keys as
strings. Currently the output format was a mix of atoms and strings.
Having a consistent key format will make payloads easier to test.
However, this may have an impact on the performance of the library and
this should be considered a breaking API change.