Deep merge relationship paths
Fixes #150

Keyword.merge was combining the nested lists instead of recursively
merging the lists, so that relationships paths that shared relationships
past the first level were never merged.  This then left the Keyword list
with multiple copies of the same key and last code line Included.build
only saw the first copy of the key (which was actually the last path in
the string format because of the prepending to the normalized
relationship path list).