Add JaSerializer.Builder.PaginationLinks
Closes #231

Introduces breaking changes in the form of the page keys.

I opted to normalize on the suggested `page[number]` and `page[size]`
nomenclature that is used as an example in the json-api `page` spec:

http://jsonapi.org/format/#fetching-pagination

Keep in mind that this nomenclature is not the suggested one, as there
isn't one but in my experience this is a very common implementation over
the original `page[page]` and `page[page-size]`

This PR also has the ScrivenerLinks builder consume `PaginationLinks`
builder.

For the time being I suggest building links and passing to the `page`
option for the view which is why this builder's API does not conform to
the same API as the other builders.

This PR also give configuration control over the page number origin
value. Previously this was hardcoded to `1` but now you can override if
you prefer the `offset/limit` nomenclature.